requirements.txt
fastapi
uvicorn
python-jose
passlib
argon2-cffi
slowapi
python-dotenv
pytest
pytest-cov
httpx
